fix: 优化代码

+++ src/main/java/com/weizujie/attendance/mapper/LeaveMapper.java
@@ -17,5 +17,5 @@ public interface LeaveMapper extends BaseMapper<Leave> {
 
     int checkLeave(Leave leave);
 
-    int deleteLeave(Long id);
+    int deleteLeave(Integer id);
 }
