+++ src/main/java/com/weizujie/attendance/utils/Data.java
@@ -3,13 +3,13 @@ package com.weizujie.attendance.utils;
 import java.util.List;
 
 public class Data {
-    private List<Long> ids ;
+    private List<Integer> ids ;
 
-    public List<Long> getIds() {
+    public List<Integer> getIds() {
         return ids;
     }
 
-    public void setIds(List<Long> ids) {
+    public void setIds(List<Integer> ids) {
         this.ids = ids;
     }
 }
